Helper

Documentation

data ConfigPlayer

Data for storing the configuration of the dice at any time

Constructors

ConfigPlayer

Fields

top :: Int

bottom :: Int

north :: Int

south :: Int

east :: Int

west :: Int

data Cell

Data representing the types of cell

Constructors

Wall

Player

Target

OnTarget

Empty

data GameState

Data representing thr game state at any time in the game

Constructors

Running

GameOver

data Turn

Player turn

Constructors

Player0

Player1

type Board = Array (Int, Int) Cell

data Game

Contains total information of the game at any given time

Constructors

Game

Fields

gameBoard :: (Board, Board)

configPlayer :: (ConfigPlayer, ConfigPlayer)

gameState :: GameState

finalTarget :: ((Int, Int), Int)

numberOfMoves :: (Int, Int)

level :: Int

isPlaying :: Turn

findAnything :: [Char] -> Int -> Char -> [Int]

Converts the string representing map into original map

findAllAnything :: [String] -> Int -> Char -> [(Int, Int)]

Converts the string representing map into original map

tupleWall :: Int -> [((Int, Int), Cell)]

make a tuple of ((Int,Int),wall)

tupleEmpty :: Int -> [((Int, Int), Cell)]

make a tuple of ((Int,Int),Empty)

tupleTarget :: Int -> [((Int, Int), Cell)]

make a tuple of ((Int,Int),Target)

tupleStart :: Int -> ([((Int, Int), Cell)], [((Int, Int), Cell)])

make a tuple of ((Int,Int),Start)

tupleOnTarget :: Int -> [((Int, Int), Cell)]

make a tuple of ((Int,Int),onTarget)

initBoard :: Int -> Board

make a tuple of ((Int,Int),Empty)

readyBoard :: Int -> (Board, Board)

for each tuple ((Int,Int),Cell) put it in a Board

initGameForMapLevel

Arguments

:: Int
0 level0 ans so on

-> Game
returns the game accordingly

Initilalises the game for the given level