Logic

Documentation

moveConfigPlayerUp :: ConfigPlayer -> ConfigPlayer

If the player choses to move the dice in north direction then congiuration changes

moveConfigPlayerDown :: ConfigPlayer -> ConfigPlayer

If the player choses to move the dice in south direction then congiuration changes

moveConfigPlayerRight :: ConfigPlayer -> ConfigPlayer

If the player choses to move the dice in east direction then congiuration changes

moveConfigPlayerLeft :: ConfigPlayer -> ConfigPlayer

If the player choses to move the dice in west direction then congiuration changes

allReachableCoords :: Board -> [(Int, Int)]

Finds all the coordinates that are reachable

findCoordsPlayer :: Board -> (Int, Int)

Finds the coordiates of the player in the given board

moveConfigPlayer :: Int -> Int -> ConfigPlayer -> ConfigPlayer

coord can be on empty, target or player

checkNewCoord :: Int -> Int -> Board -> (Int, Int) -> Bool

Checks the new coordinate of the player after the move whether it is valid or not

checkMove :: Int -> Int -> Game -> Bool

Checks the move of the player and returns the new configuration based on checknewMoves function

checkAndMove :: Int -> Int -> Game -> Game

newGameBoard :: Int -> Int -> Game -> (Board, Board)

Returns the (Board,Board) according to the move by the playing player

newConfigPlayer :: Int -> Int -> Game -> (ConfigPlayer, ConfigPlayer)

Returns the (ConfigPlayer,ConfigPlayer) according to the move by the playing player

newNumberOfMoves :: Game -> (Int, Int)

Returns the (Int,Int) updating the number of moves of the current playing player

newIsPlaying :: Game -> Turn

Returns the Turn of player who is [laying]

playerToMoveNum :: Game -> Int

Returns the number of moves of the current playing player

playerToMoveBoard :: Game -> Board

Returns the Board of the current playing player

playerToNotMoveBoard :: Game -> Board

Returns the board of the player who is not currently playing

playerToNotMoveConfig :: Game -> ConfigPlayer

Returns the config of the player who is not currently playing

movePlayer :: Game -> Key -> Game

Moves the player dice according to the key stroke provided by the user

checkVictory :: Game -> Game

Checks whether the some player is in final configuration or not

changeGame :: Int -> Game

Changes the level of the game

startNewGame :: Game -> Key -> Game

On pressing space key level is changed

transformGame :: Event -> Game -> Game

Transform the game based on the game state