Rendering

Documentation

makeTile :: Float -> Float -> Picture

This function is used for making grid and tile at any given point x,y

wallGrid :: [(Int, Int)] -> Picture

This function returns the picture that contains the wall cell

makeWall :: Int -> Int -> Color -> Picture

used to make a colored wall tile

gameGridIntial :: Int -> Picture

Takes level of the game as input and returns the picture containing tiles

cellsOfBoard :: Board -> Cell -> Picture -> Picture

Takes Board cell and picture as input and combines them and returned the combined picture

snapPictureToCell :: Picture -> (Int, Int) -> Picture

Maps each cell to the board in picture and returns the picture

cellPicPlayer :: ConfigPlayer -> Picture

Takes the configuration of the player as input and displays it on the screen in the form of picture

cellPicTarget :: Int -> Picture

colors and make the target cell

cellPicWall :: Picture

colors and make the wall cell

cellPicEmpty :: Picture

Colors the empty cell

gameGrid :: Int -> Game -> Picture

Combines all the pictures and generates the rendered image on the screen

scoreboard :: Int -> Game -> Picture

Prints the number of moves of both the plaer on the screen

makeFinal :: Game -> Picture -> Picture

Returns the picture of the player score who has won the game

whoWon :: Game -> Picture

Finds the player who won the game

playerBoard

Arguments

:: Int
o for player0 1 for player1

-> Game
game at any given time

-> Board
returned board of player

Returns the number of moves of player

playerConfigPlayer

Arguments

:: Int
o for player0 1 for player1

-> Game
game at any given time

-> ConfigPlayer
returned configuration of player

Returns the number of moves of player

playerNumOfMoves

Arguments

:: Int
o for player0 1 for player1

-> Game
game at any given time

-> Int
returned number of moves

Returns the number of moves of player

makeI :: Int -> Picture

converts given integer into picture with same number of dots

makeOne :: Picture

MAkes a Picture of one dot to represents the dice(1)

makeTwo :: Picture

MAkes a Picture of 2 dots to represents the dice(2)

makeThree :: Picture

MAkes a Picture of 3 dots to represents the dice(3)

makeFour :: Picture

MAkes a Picture of 4 dots to represents the dice(4)

makeFive :: Picture

MAkes a Picture of 5 dots to represents the dice(5)

makeSix :: Picture

MAkes a Picture of 6 dots to represents the dice(6)

gameAsPicture :: Game -> Picture

Convets the given game into picture and dislpays it in the form of the picture